evaluation vs other models/requirements.txt
faiss-cpu>=1.7.4
rank-bm25>=0.2.2
sentence-transformers>=2.2.2
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
tqdm>=4.65.0